+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.8.0] - 2022-07-04
+
+### Changed
+
+- Adding 2 new EndPoints of Campaign:
+  - Get campaign assets
+  - Get active Campaign
+
 ## [3.7.0] - 2022-06-23
 
 ### Changed